+++ src/1-one/output/methods/index.js
@@ -1,0 +1,4 @@
+import hash from './hash.js'
+import debug from './debug/index.js'
+
+export { hash, debug }