+++ src/1-one/output/plugin.js
@@ -1,11 +1,12 @@
 import api from './api/index.js'
-import hash from './api/lib/hash.js'
+import { debug, hash } from './methods/index.js'
 
 export default {
   api,
   methods: {
     one: {
-      hash
-    }
-  }
+      hash,
+      debug,
+    },
+  },
 }