+++ src/3-three/numbers/numbers/isUnit.js
@@ -1,25 +1,26 @@
 import parse from './parse/index.js'
 
-const isArray = function (arr) {
-  return Object.prototype.toString.call(arr) === '[object Array]'
-}
+const isArray = arr => Object.prototype.toString.call(arr) === '[object Array]'
 
-const toObject = function (input) {
-  if (typeof input === 'string') {
+// turn anything into {foo:true} format
+const coerceToObject = function (input) {
+  if (typeof input === 'string' || typeof input === 'number') {
     let tmp = {}
     tmp[input] = true
-    input = tmp
-  } else if (isArray(input)) {
-    input = input.reduce((h, s) => {
+    return tmp
+  }
+  if (isArray(input)) {
+    return input.reduce((h, s) => {
       h[s] = true
       return h
     }, {})
   }
-  return input
+  return input || {}
 }
 
+// only return values with the given unit
 const isUnit = function (doc, input = {}) {
-  input = toObject(input)
+  input = coerceToObject(input)
   return doc.filter(p => {
     let { unit } = parse(p)
     if (unit && input[unit] === true) {