+++ changelog.md
@@ -17,8 +17,21 @@ While all _Major_ releases should be reviewed, our only _large_ releases are **v
 
 <!-- #### 14.11.0 [Unreleased]
 
-
 -->
+
+#### 14.12.0 [Feb 2024]
+
+- **[new]** - .payload() plugin
+- **[new]** - `.numbers().isUnit()` method #1089
+- **[change]** - update github workflow (thanks FDawgs!)
+- **[fix]** - README issues (thanks track0x1!)
+- **[fix]** - .has() inconsistency
+- **[new]** - support adding debug methods via plugins
+- **[change]** - remove deprecated .debug(object) support
+
+- **[new]** - `dates().isBefore()`, `dates().isBefore()`
+- **[fix]** - join() issue
+- **[fix]** - .before() .after()
 
 #### 14.11.2 [Jan 2024]